requirements.txt
aiosqlite==0.18.0
alembic==1.9.1
anyio==3.6.2
apprise==1.2.1
asgi-lifespan==2.0.0
async-generator==1.10
asyncpg==0.27.0
attrs==22.1.0
backports.zoneinfo==0.2.1
boto3==1.26.14
botocore==1.29.14
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
commonmark==0.9.1
coolname==2.1.0
croniter==1.3.8
cryptography==39.0.0
dateparser==1.1.5
docker==6.0.1
exceptiongroup==1.0.4
fastapi==0.88.0
fsspec==2022.11.0
google-auth==2.15.0
greenlet==2.0.1
griffe==0.25.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.2
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
kubernetes==25.3.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
orjson==3.8.3
outcome==1.2.0
packaging==21.3
pathspec==0.10.3
pendulum==2.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.0
rsa==4.9
s3transfer==0.6.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.45
starlette==0.22.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.12.1
typer==0.7.0
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.20.0
webdriver-manager==3.8.5
websocket-client==1.4.2
wsproto==1.2.0
zipp==3.11.0
